+++ src/settings/TagSettings.test.js
@@ -1,4 +1,4 @@
-import { render } from '@testing-library/react';
+import { render } from '@folio/jest-config-stripes/testing-library/react';
 import { reducer as formReducer } from 'redux-form';
 import {
   createStore,